requirements.txt
certifi==2018.10.15
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.3
django-allauth==0.38.0
django-common-helpers==0.9.2
django-cors-headers==2.4.0
django-cron==0.5.1
django-filter==2.0.0
django-seed==0.1.9
django-summernote==0.8.10.0
django-tinymce==2.7.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
fake-factory==9999.9.9
Faker==1.0.0
gunicorn==19.9.0
idna==2.7
oauthlib==2.1.0
Pillow==5.3.0
PyJWT==1.6.4
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2018.7
requests==2.20.0
requests-oauthlib==1.0.0
six==1.11.0
text-unidecode==1.2
urllib3==1.24.1
whitenoise==4.1.1
